requirements.txt
gradio
google-generativeai
numpy
ipython
python-dotenv
pillow